revised README.md
The title pretty much says it all, all I did was change it a little bit. I don't know why this much of a detailed report is important but, I will still do it if the people next to me requires it. Just trying to fill up more space with this. You know how the song goes: "This is the filler, the filler ya. When you don't have enough words for the paragraph. This is the filler, filler part. Don't pay much attention; it's not important. This is the filler, the filler ya!"